pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools<80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfn-check"
version = "0.9.0"
requires-python = ">=3.12"
description = "Validate Cloud Formation"
readme = "README.md"
authors = [{ name = "Ada Lundhe", email = "adalundhe@lundhe.audio" }]
license = { file = "LICENSE" }
classifiers = [
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: 3.13",
        "Operating System :: OS Independent",
]
keywords = [
    "cloud-formation",
    "testing",
    "aws",
    "cli",
]
dependencies = [
    "pydantic",
    "ruamel.yaml",
    "hyperlight-cocoa",
    "async-logging",
]


[project.urls]
Homepage = "https://github.com/adalundhe/cfn-check"

[project.scripts]
cfn-check = "cfn_check.cli.root:run"

[tool.setuptools.packages]
find = {}  # Scanning implicit namespaces is active by default

[tool.setuptools.package-data]
'src' = ['*.json', '*.md']

[tool.ruff]
target-version = "py312"